+++ test/app/components/missing_default_collection_parameter_component.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class MissingDefaultCollectionParameterComponent < ViewComponent::Base
+  def call
+    ""
+  end
+end