+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # master
 
+* Raise error when initializer omits with_collection_parameter.
+
+    *Joel Hawksley*
+
 # 2.5.1
 
 * Compile component before rendering collection.